refactor: 네이밍 컨벤션 정리 (파일명 오타, 훅/타입 이름 통일, CardBrandRange 위치 이동)
