+++ src/constants/cardBrandImages.ts
@@ -3,7 +3,7 @@ import master from '../assets/Mastercard.png';
 import diners from '../assets/Diners.png';
 import amex from '../assets/AMEX.png';
 import unionpay from '../assets/UnionPay.png';
-import type { CardBrand } from '../types/cardStausTypes';
+import type { CardBrand } from '../types/cardStatusTypes';
 
 export const CARD_BRAND_IMAGE: Record<Exclude<CardBrand, ''>, { src: string; alt: string }> = {
   visa: {
